Fix SQL errors, wrong solutions, filename inconsistencies, and update README
- Replace wrong content in 03_einfache_joins_loesungen.sql (was subquery duplicate) with correct JOIN solutions
- Fix typo bezzeichnung -> bezeichnung in 01_select_loesungen.sql
- Fix status 'LEER' -> 'STORNIERT' to match CHECK constraint in trigger solutions
- Fix RAISE NOTICE %% escaping in 05_fehler_schleifen_cursor_loesungen.sql
- Fix CALL-in-trigger compatibility and AFTER INSERT timing in 06_mini_projekt
- Fix exercise/solution mismatches (subqueries A3, views A7, aggregation A10)
- Add missing "teuerstes Produkt je Kunde" to 10/A3, fix bind params in 10/A8
- Fix NULL handling in 02_filter_und_bedingungen A6, add NOT NULL hints for A5/A13
- Fix v_rabatt INT -> NUMERIC(5,2) and ILIKE pattern in PL/pgSQL solutions
- Fix soft hyphen, FH->Vorlesungsskript, Oracle syntax in exercise text
- Rename files for consistency (Umlaut->ASCII, cursors->cursor, Klausuren->Klausur)
- Rewrite README with correct paths, file listings, and PL/pgSQL section
- Fix folder name typo in Anleitung_Anlegen_Tabellen.md

+++ PL/SQL_Aufgaben/PL/SQL_Loesungen/01_funktionen_basic_loesungen.sql
@@ -154,7 +154,7 @@ RETURNS TEXT
 AS $$
 DECLARE
     v_name    TEXT;
-    v_rabatt  INT;
+    v_rabatt  NUMERIC(5,2);
     v_result  TEXT;
 BEGIN
     SELECT name, rabatt_prozent
